requirements.txt
smolagents[toolkit,openai,litellm,gradio]==1.17.0
scapy==2.6.1
python-dotenv==1.1.0